zmieniony typ od rozmiaru

+++ src/main/java/FileData.java
@@ -3,10 +3,10 @@ import java.util.Map;
 
 public class FileData {
     private String name;                                    //nazwa
-    private Integer size;                                   //rozmiar
+    private long size;                                   //rozmiar
     private Map<String, Map<String, Integer>> dependencies; //mapa połączeń i ich wagi
 
-    public FileData(String n,Integer i,Map<String,Map<String,Integer>> m){
+    public FileData(String n,long i,Map<String,Map<String,Integer>> m){
         name=n;
         size=i;
         dependencies=new HashMap<String,Map<String,Integer>>(m);
@@ -14,7 +14,7 @@ public class FileData {
     public String getName(){
         return name;
     }
-    public Integer getSize(){
+    public long getSize(){
         return size;
     }
     public Map<String,Map<String,Integer>> getDependencies(){
